refactor: extract Cell interface & clock system
